Pull file name from airtable metadata
This makes the fileNode pull its name from the airtable
metadata instead of the remote file, because when an airtable
user changes the file name, airtable does not rename the original
file. This change makes file name changes in airtable usable in
the fileNode instead of needing to download, rename, and
re-upload the file.